Extract method to generate random avatar
