+++ src/components/shared/PostListItem.js
@@ -8,7 +8,7 @@ import IconThumbUp from '@material-ui/icons/ThumbUp';
 import Chip from '@material-ui/core/Chip';
 import Avatar from '@material-ui/core/Avatar';
 
-import axios from '../../configs/axios';
+import { fetchUserRandomInfo } from '../../utils/randomProfile';
 
 const styles = theme => ({
   root: theme.mixins.gutters({
@@ -49,65 +49,47 @@ class PostListItem extends React.Component {
     avatarURL: '',
   };
 
-
-  componentDidMount() {
-    this.fetchUserInfo();
+  async componentDidMount() {
+    try {
+      const avatarURL = await fetchUserRandomInfo();
+      this.setState({ avatarURL });
+    } catch (e) {
+      console.log('Could not fetch user info');
+      console.log(e);
+    }
   }
-   setAvatarURL = (name) => {
-     // Red, Green and Blue
-     const COLORS = ['700', '070', '007'];
-     const avatarName = encodeURIComponent(name);
-     const ramdomIndex = Math.floor(Math.random() * 3);
-     const BASE_URL = 'https://ui-avatars.com/api/';
-     const url = `${BASE_URL}?name=${avatarName}&color=fff&background=${COLORS[ramdomIndex]}`;
-     return url;
-   }
 
-   async fetchUserInfo() {
-     // fetching anonymous name and avatar for an user
-     try {
-       const response = await axios.get('/anonymous-name/');
-       const name = response.data.anonymous_name;
-       const avatarURL = this.setAvatarURL(name);
-       this.setState({
-         avatarURL,
-       });
-     } catch (e) {
-       console.log('Could not fetch user info');
-       console.log(e);
-     }
-   }
+  render() {
+    return (
+      <Paper className={this.props.classes.root} elevation={4}>
+        <Typography variant="headline" component="h3" className={this.props.classes.typographyStyle} >
+          <Avatar src={this.state.avatarURL} />
+        </Typography>
 
-   render() {
-     return (
-       <Paper className={this.props.classes.root} elevation={4}>
-         <Typography variant="headline" component="h3" className={this.props.classes.typographyStyle} >
-           <Avatar src={this.state.avatarURL} />
-         </Typography>
+        <Typography component="p">
+          {this.props.emotion === 'g' ?
+            <IconThumbUp className={this.props.classes.goodEmotion} />
+          :
+            <IconThumbDown className={this.props.classes.badEmotion} />
+        }
+          {' '}
+          {this.props.subject}
 
-         <Typography component="p">
-           {this.props.emotion === 'g' ?
-             <IconThumbUp className={this.props.classes.goodEmotion} />
-            :
-             <IconThumbDown className={this.props.classes.badEmotion} />
-          }
-           {' '}
-           {this.props.subject}
+        </Typography>
 
-         </Typography>
+        <Typography component="p">
+          {this.props.tags.length > 0 ?
+            <React.Fragment>
+              <ListTags tags={this.props.tags} className={this.props.classes.chip} />
+            </React.Fragment>
+          :
+          null
+        }
+        </Typography>
 
-         <Typography component="p">
-           {this.props.tags.length > 0 ?
-             <React.Fragment>
-               <ListTags tags={this.props.tags} className={this.props.classes.chip} />
-             </React.Fragment>
-            :
-            null
-          }
-         </Typography>
-
-       </Paper>
-     );
-   }
+      </Paper>
+    );
+  }
 }
+
 export default withStyles(styles)(PostListItem);
